feat(rating): half-star ratings, 0–5 in 0.5 steps
Owner-requested. The scale goes to half stars everywhere it appears: the bundle
inspector, the multi-select batch bar, the toolbar Rating filter, and the Smart
Collection editor. Each star is two targets — left picks n−½, right picks n — and
clicking the current value still clears.

**The design decision that matters: a rating is stored as a number of stars, not a
count of half-star units.** So 4 stays 4 rather than becoming 8, every whole-star
rating in an existing library still reads as itself, `rating >= 4` in a saved Smart
Collection still selects four stars and up, and no migration runs — libraries move
freely between this version and the previous one. The column is REAL now; SQLite's
INTEGER affinity narrows a float only when lossless, so 4.0 stores as INTEGER 4 and
3.5 as REAL in the same column, and the existing CHECK (0..5) still holds. The step
is enforced in the service and schema layers, which is where a CHECK could not
reach without rebuilding the table.

Picking deliberately runs through each half's own `<button onClick>`. An earlier
form of this took pointer capture on `pointerdown` so it could reconstruct the pick
from the release — but capture retargets *every* later event of the gesture to the
capturing element, the click included, so the button's handler could never fire. A
real click reported `pointerdown` on the button and both `pointerup` and `click` on
the row. Chromium made that work; the desktop shell did not, and there a click on a
star did nothing at all (owner-reported). Capture is now taken on the first move
that changes value — the moment the gesture is a sweep, the only thing that needs
it — so a press-and-release never captures and its click reaches the button. A
sweep commits on release and swallows the trailing click; the suppression flag is
cleared when a fresh gesture starts, because a sweep released outside the row
dispatches no click to consume it and a stale flag silently ate the next pick.

Presentation: a hover or sweep previews the fill and names the value in a hint; the
fill is a second ★ clipped over a muted one, so the half lands on the glyph's true
midpoint in any font rather than depending on ☆ and ★ sharing an advance width. The
filter's per-star count follows whichever half is under the pointer, so it always
answers "how many would clicking here match?". Facet keys gain halves ("3.5") while
whole stars keep the key they had ("4", never "4.0"). Sizes were tuned by the owner
to 18px in the inspector and 24px in the filter popover, with tighter spacing in
both filter rows.

Verified with a real mouse against the server rather than the optimistic UI: a
click on "3½ stars" stores 3.5, a sweep from "1 star" to "4½ stars" stores 4.5.
Gates: backend 896 tests, web 488, browser e2e 94 passed (the one failure is the
pre-existing HLS re-attach case, which fails on main too).

**Not verified on WebKit** — it cannot be driven from here. The fix removes the
dependency on capture semantics rather than working around a quirk, which is the
reason to expect it to hold; the owner's retest in the desktop shell is the
confirmation. The inspector's half targets are 9px at the chosen size, which is the
number to revisit if picking ever feels fiddly again.

+++ apps/server/src/cairndex/persistence/models.py
@@ -130,8 +130,13 @@ class AssetBundle(Base):
     # no predefined roles; each entry is just a separate text block. NULL on rows
     # created before this column existed reads back as an empty list.
     notes: Mapped[list[str] | None] = mapped_column(JSON, nullable=True)
-    # 0..5; NULL means unrated. Range enforced by a CHECK constraint and schema.
-    rating: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # 0..5 stars in half-star steps; NULL means unrated (see domain/rating.py).
+    # The CHECK below enforces the range. The *step* is enforced in the service
+    # and schema layers only: a library created before half stars has this table
+    # already built, and SQLite cannot add a constraint to an existing table
+    # without rebuilding it — which is exactly what storing stars rather than
+    # half-units lets us avoid.
+    rating: Mapped[float | None] = mapped_column(Float, nullable=True)
 
     # The selected cover lives in this bundle. The FK forms a cycle with
     # asset_files.bundle_id, so it is created via ALTER and updated after the
